style the login page
